requirements.txt
streamlit>=1.22
# Upgrade OpenAI for new Responses API & modern client
openai>=1.0.0

# Retain legacy langchain for backward compatibility (optional)
langchain>=0.0.380
# Add split packages for modern LangChain layout
langchain-community>=0.2.0
langchain-openai>=0.2.0
langchain-experimental>=0.0.55

# Multi‑provider LLM support (Anthropic / Gemini / LLaMA / optional Groq)
anthropic>=0.34.0
google-generativeai>=0.8.0
huggingface-hub>=0.23.0
# groq SDK optional (we currently use raw HTTP; install if switching to SDK)
groq>=0.5.0

snowflake-connector-python>=3.0.3
python-dotenv>=1.0.0
pandas>=2.0.0
sqlparse>=0.5.3
tenacity>=8.2.2
altair>=4.2.0
python-json-logger>=2.0.4
requests>=2.31.0